requirements.txt
# Azure Functions Core
azure-functions==1.18.0

# Azure SDK packages (reduced since we're using output bindings for Cosmos DB)
azure-storage-blob==12.19.0

# HTTP requests for AI Content Understanding
requests==2.31.0

# Additional packages for document processing
# You can add more packages based on your specific needs:
# PyPDF2==3.0.1          # For PDF processing
# python-docx==1.1.0     # For Word documents
# openpyxl==3.1.2        # For Excel files
# Pillow==10.2.0         # For image processing
# opencv-python==4.8.1.78  # For advanced image analysis
